Move the diff library into pkg for use by external tooling
This will be helpful in zed for doing diffing between schemas
