+++ pkg/diff/caveats/diff_test.go
