+++ pkg/diff/namespace/diff_test.go
